src/generator/requirements.txt
# Usage: pip install -r requirements.txt
Pillow>=7.1.2
